Main

+++ Backend/Controllers/TestCase/testStep.js
@@ -15,10 +15,19 @@ const saveTestStep = async (req, res) => {
   */
 
   try {
+    const { processId, reusableProcessId, step } = req.body;
+
+    let permissionName = reusableProcessId ? "Reusable Process" : "Test Case";
+
+    if (!req.user.customerAdmin) {
+      const allowed = await req.user.permissions.some((permission) => {
+        return permissionName == permission.name && permission.edit;
+      });
+      if (!allowed) return res.status(401).json({ error: "Unauthorized" });
+    }
+
     const { error } = saveTestStepValidation.validate(req.body);
     if (error) throw new Error(error.details[0].message);
-
-    const { processId, reusableProcessId, step } = req.body;
 
     if (processId) {
       await TestStep.schema(req.database).increment("step", {
@@ -66,6 +75,22 @@ const updateTestStep = async (req, res) => {
 
   try {
     const testStepId = req.params.testStepId;
+
+    const updatingStep = await TestStep.schema(req.database).findByPk(
+      testStepId
+    );
+
+    let permissionName = updatingStep.reusableProcessId
+      ? "Reusable Process"
+      : "Test Case";
+
+    if (!req.user.customerAdmin) {
+      const allowed = await req.user.permissions.some((permission) => {
+        return permissionName == permission.name && permission.edit;
+      });
+      if (!allowed) return res.status(401).json({ error: "Unauthorized" });
+    }
+
     const { error } = updateTestStepValidation.validate({
       ...req.body,
       testStepId,
@@ -119,11 +144,24 @@ const deleteTestStep = async (req, res) => {
 
   try {
     const testStepId = req.params.testStepId;
-    const { error } = idValidation.validate({ id: testStepId });
-    if (error) throw new Error(error.details[0].message);
+
     const deletingTestStep = await TestStep.schema(req.database).findByPk(
       testStepId
     );
+
+    let permissionName = deletingTestStep.reusableProcessId
+      ? "Reusable Process"
+      : "Test Case";
+
+    if (!req.user.customerAdmin) {
+      const allowed = await req.user.permissions.some((permission) => {
+        return permissionName == permission.name && permission.edit;
+      });
+      if (!allowed) return res.status(401).json({ error: "Unauthorized" });
+    }
+
+    const { error } = idValidation.validate({ id: testStepId });
+    if (error) throw new Error(error.details[0].message);
 
     const deletedTestStep = await TestStep.schema(req.database).destroy({
       where: { id: testStepId },
